update command server
